--- a/src/compute/Cargo.toml
+++ b/src/compute/Cargo.toml
@@ -7,7 +7,7 @@ rust-version.workspace = true
 publish = false
 
 [dependencies]
-ahash = "0.8.0"
+ahash = { version = "0.8.0", default_features = false }
 anyhow = "1.0.66"
 async-trait = "0.1.68"
 bytesize = "1.1.0"