--- a/src/compute/src/extensions/collection.rs
+++ b/src/compute/src/extensions/collection.rs
@@ -60,26 +60,7 @@ where
         Arranged<G, TraceAgent<Tr>>: ArrangementSize,
     {
         if must_consolidate {
-            // We employ AHash below instead of the default hasher in DD to obtain
-            // a better distribution of data to workers. AHash claims empirically
-            // both speed and high quality, according to
-            // https://github.com/tkaitchuck/aHash/blob/master/compare/readme.md.
-            // TODO(vmarcos): Consider here if it is worth it to spend the time to
-            // implement twisted tabulation hashing as proposed in Mihai Patrascu,
-            // Mikkel Thorup: Twisted Tabulation Hashing. SODA 2013: 209-228, available
-            // at https://epubs.siam.org/doi/epdf/10.1137/1.9781611973105.16. The latter
-            // would provide good bounds for balls-into-bins problems when the number of
-            // bins is small (as is our case), so we'd have a theoretical guarantee.
-            let random_state = ahash::RandomState::new();
-            let mut h = random_state.build_hasher();
-            let exchange = Exchange::new(move |update: &((D1, _), G::Timestamp, R)| {
-                let data = &(update.0).0;
-                data.hash(&mut h);
-                h.finish()
-            });
-            self.map(|k| (k, ()))
-                .mz_arrange_core::<_, Tr>(exchange, name)
-                .as_collection(|d: &D1, _| d.clone())
+            self.mz_consolidate(name)
         } else {
             self.clone()
         }
@@ -91,8 +72,35 @@ where
         Tr::Batch: Batch,
         Arranged<G, TraceAgent<Tr>>: ArrangementSize,
     {
+        // We employ AHash below instead of the default hasher in DD to obtain
+        // a better distribution of data to workers. AHash claims empirically
+        // both speed and high quality, according to
+        // https://github.com/tkaitchuck/aHash/blob/master/compare/readme.md.
+        // TODO(vmarcos): Consider here if it is worth it to spend the time to
+        // implement twisted tabulation hashing as proposed in Mihai Patrascu,
+        // Mikkel Thorup: Twisted Tabulation Hashing. SODA 2013: 209-228, available
+        // at https://epubs.siam.org/doi/epdf/10.1137/1.9781611973105.16. The latter
+        // would provide good bounds for balls-into-bins problems when the number of
+        // bins is small (as is our case), so we'd have a theoretical guarantee.
+        // NOTE: We fix the seeds of a RandomState instance explicity with the same
+        // seeds that would be given by `AHash` via ahash::AHasher::default() so as
+        // to avoid a different selection due to compilation flags being differently
+        // selected in other dependencies using `AHash` vis-à-vis cargo's strategy
+        // of unioning compilation flags.
+        let random_state = ahash::RandomState::with_seeds(
+            0x243f_6a88_85a3_08d3,
+            0x1319_8a2e_0370_7344,
+            0xa409_3822_299f_31d0,
+            0x082e_fa98_ec4e_6c89,
+        );
+        let exchange = Exchange::new(move |update: &((D1, _), G::Timestamp, R)| {
+            let data = &(update.0).0;
+            let mut h = random_state.build_hasher();
+            data.hash(&mut h);
+            h.finish()
+        });
         self.map(|k| (k, ()))
-            .mz_arrange::<Tr>(name)
+            .mz_arrange_core::<_, Tr>(exchange, name)
             .as_collection(|d: &D1, _| d.clone())
     }
 }