--- a/test/cluster/mzcompose.py
+++ b/test/cluster/mzcompose.py
@@ -63,6 +63,7 @@ def workflow_default(c: Composition, parser: WorkflowArgumentParser) -> None:
         "test-github-17177",
         "test-github-17510",
         "test-github-17509",
+        "test-github-19610",
         "test-remote-storage",
         "test-drop-default-cluster",
         "test-upsert",
@@ -971,6 +972,118 @@ def workflow_test_github_17509(c: Composition) -> None:
         c1 = c.invoke("logs", "clusterd_nopanic", capture=True)
         assert "Non-positive accumulation in MinsMaxesHierarchical" in c1.stdout
         assert "Negative accumulation in ReduceMinsMaxes" not in c1.stdout
+
+
+def workflow_test_github_19610(c: Composition) -> None:
+    """
+    Test that a monotonic one-shot SELECT will perform consolidation without error on valid data.
+    We introduce data that results in a multiset and compute min/max. In a monotonic one-shot
+    evaluation strategy, we must consolidate and subsequently assert monotonicity.
+
+    This is a regression test for https://github.com/MaterializeInc/materialize/issues/19610, where
+    we observed a performance regression caused by a correctness issue. Here, we validate that the
+    underlying correctness issue has been fixed.
+    """
+
+    c.down(destroy_volumes=True)
+    with c.override(
+        Clusterd(
+            name="clusterd_nopanic",
+            environment_extra=[
+                "MZ_PERSIST_COMPACTION_DISABLED=true",
+            ],
+        ),
+        Testdrive(no_reset=True),
+    ):
+        c.up("testdrive", persistent=True)
+        c.up("materialized")
+        c.up("clusterd_nopanic")
+
+        c.sql(
+            "ALTER SYSTEM SET enable_unmanaged_cluster_replicas = true;",
+            port=6877,
+            user="mz_system",
+        )
+
+        c.sql(
+            "ALTER SYSTEM SET enable_repeat_row = true;",
+            port=6877,
+            user="mz_system",
+        )
+
+        c.sql(
+            "ALTER SYSTEM SET enable_monotonic_oneshot_selects = true;",
+            port=6877,
+            user="mz_system",
+        )
+
+        # set up a test cluster and run a testdrive regression script
+        c.sql(
+            """
+            CREATE CLUSTER cluster1 REPLICAS (
+                r1 (
+                    STORAGECTL ADDRESSES ['clusterd_nopanic:2100'],
+                    STORAGE ADDRESSES ['clusterd_nopanic:2103'],
+                    COMPUTECTL ADDRESSES ['clusterd_nopanic:2101'],
+                    COMPUTE ADDRESSES ['clusterd_nopanic:2102'],
+                    WORKERS 4
+                )
+            );
+            -- Set data for test up.
+            SET cluster = cluster1;
+            CREATE TABLE base (data bigint, diff bigint);
+            CREATE MATERIALIZED VIEW data AS SELECT data FROM base, repeat_row(diff);
+            INSERT INTO base VALUES (1, 6);
+            INSERT INTO base VALUES (1, -3), (1, -2);
+            INSERT INTO base VALUES (2, 3), (2, 2);
+            INSERT INTO base VALUES (2, -1), (2, -1);
+            INSERT INTO base VALUES (3, 3), (3, 2);
+            INSERT INTO base VALUES (3, -3), (3, -2);
+            INSERT INTO base VALUES (4, 1), (4, 2);
+            INSERT INTO base VALUES (4, -1), (4, -2);
+            INSERT INTO base VALUES (5, 5), (5, 6);
+            INSERT INTO base VALUES (5, -5), (5, -6);
+            """
+        )
+        c.testdrive(
+            dedent(
+                """
+            > SET cluster = cluster1;
+
+            # Computing min/max with a monotonic one-shot SELECT requires
+            # consolidation. We test here that consolidation works correctly,
+            # since we assert monotonicity right after consolidating.
+            # Note that we employ a cursor to avoid testdrive retries.
+            # Hash functions used for exchanges in consolidation may be
+            # nondeterministic and produce the correct output by chance.
+            > BEGIN
+            > DECLARE cur CURSOR FOR SELECT min(data), max(data) FROM data;
+            > FETCH ALL cur;
+            1 2
+            > COMMIT;
+
+            # To reduce the chance of a (un)lucky strike of the hash function,
+            # let's do the same a few times.
+            > BEGIN
+            > DECLARE cur CURSOR FOR SELECT min(data), max(data) FROM data;
+            > FETCH ALL cur;
+            1 2
+            > COMMIT;
+
+            > BEGIN
+            > DECLARE cur CURSOR FOR SELECT min(data), max(data) FROM data;
+            > FETCH ALL cur;
+            1 2
+            > COMMIT;
+
+            > BEGIN
+            > DECLARE cur CURSOR FOR SELECT min(data), max(data) FROM data;
+            > FETCH ALL cur;
+            1 2
+            > COMMIT;
+            """
+            )
+        )
 
 
 def workflow_test_upsert(c: Composition) -> None: