This PR fixes a correctness bug in the usage of the hashing API, in particular `AHash`'s, in the `Exchange` pact of `mz_consolidate_if`. Additionally, the commit extends the patched usage of the hash function to `mz_consolidate`, thus increasing our confidence on good distribution of rows to workers given the use of a higher quality hash function than the DD's default `FnvHasher`. The bug was caused by: (a) initializing the hash function with a method that would employ randomness, thus leading to divergent hash state across workers; (b) not reinitializing, but reusing the hash function across invocations, thus computing from a hash state that was not well-defined.

Fixes #19610.

### Motivation

  * This PR fixes a recognized bug. https://github.com/MaterializeInc/materialize/issues/19610

### Tips for reviewer

The method `ahash::AHasher::default()` should construct an instance of the hash function with fixed seeding, but this depends on the `cfg` attributes given. Since `AHash` is used extensively, relying on such attributes may be fragile, so I tried to employ a method that is unconditionally seeding the input `RandomState` with fixed seeds (see [comments](https://github.com/MaterializeInc/materialize/pull/19848#issuecomment-1591259633) [below](https://github.com/MaterializeInc/materialize/pull/19848#issuecomment-1592722178)). By reading the code of `AHash`, this usage seems to imply that we end up with the fallback hasher, unfortunately. However, this hash function [should be sufficient](https://github.com/tkaitchuck/aHash/blob/master/FAQ.md#how-is-ahash-so-fast) for our requirements.

Additionally, as per the reasons articulated above, the call to obtain a new hash function instance from this fixed `RandomState` must be performed at each invocation of the `Exchange` closure. Note that this is exactly the same usage pattern of the hash function employed in DD's code with `FnvHasher`, except that in the case of the `FnvHasher` the `Default::default()` implementation has unconditionally assigned fixed seeds.

### Checklist

- [x] This PR has adequate test coverage / QA involvement has been duly considered.
- [X] This PR has an associated up-to-date [design doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/20230421_stabilize_monotonic_select.md).
- [X] If this PR evolves [an existing `$T ⇔ Proto$T` mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md) (possibly in a backwards-incompatible way), then it is tagged with a `T-proto` label. N/A
- [X] If this PR will require changes to cloud orchestration, there is a companion cloud PR to account for those changes that is tagged with the release-blocker label ([example](https://github.com/MaterializeInc/cloud/pull/5021)). N/A
- [X] This PR includes the following [user-facing behavior changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note): N/A