@vmarcos 2023-06-14
I rebased this PR on main and observed the following locally with the repro given in #19610:

```
This PR, 1 worker: Time: Time: 12423.220 ms (00:12.423)
Main, 1 worker: Time: 12569.347 ms (00:12.569)

This PR, 8 workers: Time: 2683.078 ms (00:02.683)
Main, 8 workers: Time: 2622.184 ms (00:02.622)

With AHash, 4-4 workers: Time: 4698.016 ms (00:04.698)
Main, 4-4 workers: Time: 3067.252 ms (00:03.067)
```

Note that the 4-4 configuration oversubscribes my hardware, so it runs a bit slower than the 8 worker configuration on main. The performance degradation with the PR occurs in this configuration, even after taking the measures in commit 2f03e330bd7df28193e3c1fb0060bcf80685740e. The output of the memory visualizer matches the correctness problem observed previously in a single-process setting, but now only occurring in multi-process setups:

```
This PR, dataflow records (4-4): 29119307
Main, dataflow records (4-4): 23463545
```

So we still need to ensure determinism of the hash function in multi-process settings.