@vmarcos 2023-06-15
The more reliable seeding method introduced in 1e1a9b99397e7f04c1a7521b0f1cfff63127c039 solves the issue for multi-process replicas. As expected, we can see that [a run of the feature benchmark](https://buildkite.com/materialize/nightlies/builds/2600#0188babd-00aa-4d3e-bbde-e62519face45) on this commit confirms that the regression is resolved.